Writable stream updates
Passes most tests except blocked on https://github.com/jsdom/webidl2js/pull/123#issuecomment-614278671 and https://github.com/jsdom/webidl2js/issues/79
