+++ reference-implementation/lib/QueuingStrategy.webidl
@@ -1,0 +1,6 @@
+dictionary QueuingStrategy {
+  unrestricted double highWaterMark;
+  QueuingStrategySize size;
+};
+
+callback QueuingStrategySize = unrestricted double (optional any chunk);
