+++ reference-implementation/lib/WritableStreamDefaultController.webidl
@@ -1,0 +1,4 @@
+[Exposed=(Window,Worker,Worklet)]
+interface WritableStreamDefaultController {
+  void error(optional any e);
+};
