+++ reference-implementation/lib/abstract-ops/internal-methods.js
@@ -1,0 +1,4 @@
+'use strict';
+
+exports.AbortSteps = Symbol('[[AbortSteps]]');
+exports.ErrorSteps = Symbol('[[ErrorSteps]]');
