+++ reference-implementation/lib/abstract-ops/miscellaneous.js
@@ -1,0 +1,17 @@
+'use strict';
+
+exports.IsNonNegativeNumber = v => {
+  if (typeof v !== 'number') {
+    return false;
+  }
+
+  if (Number.isNaN(v)) {
+    return false;
+  }
+
+  if (v < 0) {
+    return false;
+  }
+
+  return true;
+};
