+++ reference-implementation/lib/index.js
@@ -2,17 +2,15 @@
 // This file is used as the entry point for browserifying the reference implementation to allow it
 // to run inside the wpt-runner "browser like" context.
 
-const { ReadableStream } = require('./readable-stream.js');
-const { WritableStream } = require('./writable-stream.js');
-const { TransformStream } = require('./transform-stream.js');
-const ByteLengthQueuingStrategy = require('../generated/ByteLengthQueuingStrategy.js');
-const CountQueuingStrategy = require('../generated/CountQueuingStrategy.js');
+// const { ReadableStream } = require('./readable-stream.js');
+// const { TransformStream } = require('./transform-stream.js');
 
-window.ReadableStream = ReadableStream;
-window.WritableStream = WritableStream;
-window.TransformStream = TransformStream;
-window.CountQueuingStrategy = CountQueuingStrategy;
+// window.ReadableStream = ReadableStream;
+// window.TransformStream = TransformStream;
 window.gc = gc;
 
-ByteLengthQueuingStrategy.install(window);
-CountQueuingStrategy.install(window);
+require('../generated/ByteLengthQueuingStrategy.js').install(window);
+require('../generated/CountQueuingStrategy.js').install(window);
+require('../generated/WritableStream.js').install(window);
+require('../generated/WritableStreamDefaultController.js').install(window);
+require('../generated/WritableStreamDefaultWriter.js').install(window);
