--- reference-implementation/lib/writable-stream.js
@@ -1,1114 +1,0 @@
-'use strict';
-const assert = require('assert');
-
-// Calls to verbose() are purely for debugging the reference implementation and tests. They are not part of the standard
-// and do not appear in the standard text.
-const verbose = require('debug')('streams:writable-stream:verbose');
-
-const { CreateAlgorithmFromUnderlyingMethod, InvokeOrNoop, ValidateAndNormalizeHighWaterMark, IsNonNegativeNumber,
-        MakeSizeAlgorithmFromSizeFunction, typeIsObject, newPromise, promiseResolvedWith, promiseRejectedWith,
-        uponPromise, setPromiseIsHandledToTrue } = require('./helpers.js');
-const { DequeueValue, EnqueueValueWithSize, PeekQueueValue, ResetQueue } = require('./queue-with-sizes.js');
-
-const AbortSteps = Symbol('[[AbortSteps]]');
-const ErrorSteps = Symbol('[[ErrorSteps]]');
-
-class WritableStream {
-  constructor(underlyingSink = {}, strategy = {}) {
-    InitializeWritableStream(this);
-
-    const size = strategy.size;
-    let highWaterMark = strategy.highWaterMark;
-
-    const type = underlyingSink.type;
-
-    if (type !== undefined) {
-      throw new RangeError('Invalid type is specified');
-    }
-
-    const sizeAlgorithm = MakeSizeAlgorithmFromSizeFunction(size);
-    if (highWaterMark === undefined) {
-      highWaterMark = 1;
-    }
-    highWaterMark = ValidateAndNormalizeHighWaterMark(highWaterMark);
-
-    SetUpWritableStreamDefaultControllerFromUnderlyingSink(this, underlyingSink, highWaterMark, sizeAlgorithm);
-  }
-
-  get locked() {
-    if (IsWritableStream(this) === false) {
-      throw streamBrandCheckException('locked');
-    }
-
-    return IsWritableStreamLocked(this);
-  }
-
-  abort(reason) {
-    if (IsWritableStream(this) === false) {
-      return promiseRejectedWith(streamBrandCheckException('abort'));
-    }
-
-    if (IsWritableStreamLocked(this) === true) {
-      return promiseRejectedWith(new TypeError('Cannot abort a stream that already has a writer'));
-    }
-
-    return WritableStreamAbort(this, reason);
-  }
-
-  close() {
-    if (IsWritableStream(this) === false) {
-      return promiseRejectedWith(streamBrandCheckException('close'));
-    }
-
-    if (IsWritableStreamLocked(this) === true) {
-      return promiseRejectedWith(new TypeError('Cannot close a stream that already has a writer'));
-    }
-
-    if (WritableStreamCloseQueuedOrInFlight(this) === true) {
-      return promiseRejectedWith(new TypeError('Cannot close an already-closing stream'));
-    }
-
-    return WritableStreamClose(this);
-  }
-
-  getWriter() {
-    if (IsWritableStream(this) === false) {
-      throw streamBrandCheckException('getWriter');
-    }
-
-    return AcquireWritableStreamDefaultWriter(this);
-  }
-}
-
-module.exports = {
-  AcquireWritableStreamDefaultWriter,
-  CreateWritableStream,
-  IsWritableStream,
-  IsWritableStreamLocked,
-  WritableStream,
-  WritableStreamAbort,
-  WritableStreamClose,
-  WritableStreamDefaultControllerErrorIfNeeded,
-  WritableStreamDefaultWriterCloseWithErrorPropagation,
-  WritableStreamDefaultWriterRelease,
-  WritableStreamDefaultWriterWrite,
-  WritableStreamCloseQueuedOrInFlight
-};
-
-// Abstract operations for the WritableStream.
-
-function AcquireWritableStreamDefaultWriter(stream) {
-  return new WritableStreamDefaultWriter(stream);
-}
-
-// Throws if and only if startAlgorithm throws.
-function CreateWritableStream(startAlgorithm, writeAlgorithm, closeAlgorithm, abortAlgorithm, highWaterMark = 1,
-                              sizeAlgorithm = () => 1) {
-  assert(IsNonNegativeNumber(highWaterMark) === true);
-
-  const stream = Object.create(WritableStream.prototype);
-  InitializeWritableStream(stream);
-
-  const controller = Object.create(WritableStreamDefaultController.prototype);
-
-  SetUpWritableStreamDefaultController(stream, controller, startAlgorithm, writeAlgorithm, closeAlgorithm,
-                                       abortAlgorithm, highWaterMark, sizeAlgorithm);
-  return stream;
-}
-
-function InitializeWritableStream(stream) {
-  stream._state = 'writable';
-
-  // The error that will be reported by new method calls once the state becomes errored. Only set when [[state]] is
-  // 'erroring' or 'errored'. May be set to an undefined value.
-  stream._storedError = undefined;
-
-  stream._writer = undefined;
-
-  // Initialize to undefined first because the constructor of the controller checks this
-  // variable to validate the caller.
-  stream._writableStreamController = undefined;
-
-  // This queue is placed here instead of the writer class in order to allow for passing a writer to the next data
-  // producer without waiting for the queued writes to finish.
-  stream._writeRequests = [];
-
-  // Write requests are removed from _writeRequests when write() is called on the underlying sink. This prevents
-  // them from being erroneously rejected on error. If a write() call is in-flight, the request is stored here.
-  stream._inFlightWriteRequest = undefined;
-
-  // The promise that was returned from writer.close(). Stored here because it may be fulfilled after the writer
-  // has been detached.
-  stream._closeRequest = undefined;
-
-  // Close request is removed from _closeRequest when close() is called on the underlying sink. This prevents it
-  // from being erroneously rejected on error. If a close() call is in-flight, the request is stored here.
-  stream._inFlightCloseRequest = undefined;
-
-  // The promise that was returned from writer.abort(). This may also be fulfilled after the writer has detached.
-  stream._pendingAbortRequest = undefined;
-
-  // The backpressure signal set by the controller.
-  stream._backpressure = false;
-}
-
-function IsWritableStream(x) {
-  if (!typeIsObject(x)) {
-    return false;
-  }
-
-  if (!Object.prototype.hasOwnProperty.call(x, '_writableStreamController')) {
-    return false;
-  }
-
-  return true;
-}
-
-function IsWritableStreamLocked(stream) {
-  assert(IsWritableStream(stream) === true);
-
-  if (stream._writer === undefined) {
-    return false;
-  }
-
-  return true;
-}
-
-function WritableStreamAbort(stream, reason) {
-  const state = stream._state;
-  if (state === 'closed' || state === 'errored') {
-    return promiseResolvedWith(undefined);
-  }
-  if (stream._pendingAbortRequest !== undefined) {
-    return stream._pendingAbortRequest._promise;
-  }
-
-  assert(state === 'writable' || state === 'erroring');
-
-  let wasAlreadyErroring = false;
-  if (state === 'erroring') {
-    wasAlreadyErroring = true;
-    // reason will not be used, so don't keep a reference to it.
-    reason = undefined;
-  }
-
-  const promise = newPromise((resolve, reject) => {
-    stream._pendingAbortRequest = {
-      _resolve: resolve,
-      _reject: reject,
-      _reason: reason,
-      _wasAlreadyErroring: wasAlreadyErroring
-    };
-  });
-  stream._pendingAbortRequest._promise = promise;
-
-  if (wasAlreadyErroring === false) {
-    WritableStreamStartErroring(stream, reason);
-  }
-
-  return promise;
-}
-
-function WritableStreamClose(stream) {
-  const state = stream._state;
-  if (state === 'closed' || state === 'errored') {
-    return promiseRejectedWith(new TypeError(
-      `The stream (in ${state} state) is not in the writable state and cannot be closed`));
-  }
-
-  assert(state === 'writable' || state === 'erroring');
-  assert(WritableStreamCloseQueuedOrInFlight(stream) === false);
-
-  const promise = newPromise((resolve, reject) => {
-    const closeRequest = {
-      _resolve: resolve,
-      _reject: reject
-    };
-
-    stream._closeRequest = closeRequest;
-  });
-
-  const writer = stream._writer;
-  if (writer !== undefined && stream._backpressure === true && state === 'writable') {
-    defaultWriterReadyPromiseResolve(writer);
-  }
-
-  WritableStreamDefaultControllerClose(stream._writableStreamController);
-
-  return promise;
-}
-
-// WritableStream API exposed for controllers.
-
-function WritableStreamAddWriteRequest(stream) {
-  assert(IsWritableStreamLocked(stream) === true);
-  assert(stream._state === 'writable');
-
-  const promise = newPromise((resolve, reject) => {
-    const writeRequest = {
-      _resolve: resolve,
-      _reject: reject
-    };
-
-    stream._writeRequests.push(writeRequest);
-  });
-
-  return promise;
-}
-
-function WritableStreamDealWithRejection(stream, error) {
-  verbose('WritableStreamDealWithRejection(stream, %o)', error);
-  const state = stream._state;
-
-  if (state === 'writable') {
-    WritableStreamStartErroring(stream, error);
-    return;
-  }
-
-  assert(state === 'erroring');
-  WritableStreamFinishErroring(stream);
-}
-
-function WritableStreamStartErroring(stream, reason) {
-  verbose('WritableStreamStartErroring(stream, %o)', reason);
-  assert(stream._storedError === undefined);
-  assert(stream._state === 'writable');
-
-  const controller = stream._writableStreamController;
-  assert(controller !== undefined);
-
-  stream._state = 'erroring';
-  stream._storedError = reason;
-  const writer = stream._writer;
-  if (writer !== undefined) {
-    WritableStreamDefaultWriterEnsureReadyPromiseRejected(writer, reason);
-  }
-
-  if (WritableStreamHasOperationMarkedInFlight(stream) === false && controller._started === true) {
-    WritableStreamFinishErroring(stream);
-  }
-}
-
-function WritableStreamFinishErroring(stream) {
-  verbose('WritableStreamFinishErroring()');
-  assert(stream._state === 'erroring');
-  assert(WritableStreamHasOperationMarkedInFlight(stream) === false);
-  stream._state = 'errored';
-  stream._writableStreamController[ErrorSteps]();
-
-  const storedError = stream._storedError;
-  for (const writeRequest of stream._writeRequests) {
-    writeRequest._reject(storedError);
-  }
-  stream._writeRequests = [];
-
-  if (stream._pendingAbortRequest === undefined) {
-    WritableStreamRejectCloseAndClosedPromiseIfNeeded(stream);
-    return;
-  }
-
-  const abortRequest = stream._pendingAbortRequest;
-  stream._pendingAbortRequest = undefined;
-
-  if (abortRequest._wasAlreadyErroring === true) {
-    abortRequest._reject(storedError);
-    WritableStreamRejectCloseAndClosedPromiseIfNeeded(stream);
-    return;
-  }
-
-  const promise = stream._writableStreamController[AbortSteps](abortRequest._reason);
-  uponPromise(
-    promise,
-    () => {
-      abortRequest._resolve();
-      WritableStreamRejectCloseAndClosedPromiseIfNeeded(stream);
-    },
-    reason => {
-      abortRequest._reject(reason);
-      WritableStreamRejectCloseAndClosedPromiseIfNeeded(stream);
-    });
-}
-
-function WritableStreamFinishInFlightWrite(stream) {
-  assert(stream._inFlightWriteRequest !== undefined);
-  stream._inFlightWriteRequest._resolve(undefined);
-  stream._inFlightWriteRequest = undefined;
-}
-
-function WritableStreamFinishInFlightWriteWithError(stream, error) {
-  assert(stream._inFlightWriteRequest !== undefined);
-  stream._inFlightWriteRequest._reject(error);
-  stream._inFlightWriteRequest = undefined;
-
-  assert(stream._state === 'writable' || stream._state === 'erroring');
-
-  WritableStreamDealWithRejection(stream, error);
-}
-
-function WritableStreamFinishInFlightClose(stream) {
-  assert(stream._inFlightCloseRequest !== undefined);
-  stream._inFlightCloseRequest._resolve(undefined);
-  stream._inFlightCloseRequest = undefined;
-
-  const state = stream._state;
-
-  assert(state === 'writable' || state === 'erroring');
-
-  if (state === 'erroring') {
-    // The error was too late to do anything, so it is ignored.
-    stream._storedError = undefined;
-    if (stream._pendingAbortRequest !== undefined) {
-      stream._pendingAbortRequest._resolve();
-      stream._pendingAbortRequest = undefined;
-    }
-  }
-
-  stream._state = 'closed';
-
-  const writer = stream._writer;
-  if (writer !== undefined) {
-    defaultWriterClosedPromiseResolve(writer);
-  }
-
-  assert(stream._pendingAbortRequest === undefined);
-  assert(stream._storedError === undefined);
-}
-
-function WritableStreamFinishInFlightCloseWithError(stream, error) {
-  assert(stream._inFlightCloseRequest !== undefined);
-  stream._inFlightCloseRequest._reject(error);
-  stream._inFlightCloseRequest = undefined;
-
-  assert(stream._state === 'writable' || stream._state === 'erroring');
-
-  // Never execute sink abort() after sink close().
-  if (stream._pendingAbortRequest !== undefined) {
-    stream._pendingAbortRequest._reject(error);
-    stream._pendingAbortRequest = undefined;
-  }
-  WritableStreamDealWithRejection(stream, error);
-}
-
-// TODO(ricea): Fix alphabetical order.
-function WritableStreamCloseQueuedOrInFlight(stream) {
-  if (stream._closeRequest === undefined && stream._inFlightCloseRequest === undefined) {
-    return false;
-  }
-
-  return true;
-}
-
-function WritableStreamHasOperationMarkedInFlight(stream) {
-  if (stream._inFlightWriteRequest === undefined && stream._inFlightCloseRequest === undefined) {
-    verbose('WritableStreamHasOperationMarkedInFlight() is false');
-    return false;
-  }
-
-  verbose('WritableStreamHasOperationMarkedInFlight() is true');
-  return true;
-}
-
-function WritableStreamMarkCloseRequestInFlight(stream) {
-  assert(stream._inFlightCloseRequest === undefined);
-  assert(stream._closeRequest !== undefined);
-  stream._inFlightCloseRequest = stream._closeRequest;
-  stream._closeRequest = undefined;
-}
-
-function WritableStreamMarkFirstWriteRequestInFlight(stream) {
-  assert(stream._inFlightWriteRequest === undefined);
-  assert(stream._writeRequests.length !== 0);
-  stream._inFlightWriteRequest = stream._writeRequests.shift();
-}
-
-function WritableStreamRejectCloseAndClosedPromiseIfNeeded(stream) {
-  verbose('WritableStreamRejectCloseAndClosedPromiseIfNeeded()');
-  assert(stream._state === 'errored');
-  if (stream._closeRequest !== undefined) {
-    assert(stream._inFlightCloseRequest === undefined);
-
-    stream._closeRequest._reject(stream._storedError);
-    stream._closeRequest = undefined;
-  }
-  const writer = stream._writer;
-  if (writer !== undefined) {
-    defaultWriterClosedPromiseReject(writer, stream._storedError);
-    setPromiseIsHandledToTrue(writer._closedPromise);
-  }
-}
-
-function WritableStreamUpdateBackpressure(stream, backpressure) {
-  assert(stream._state === 'writable');
-  assert(WritableStreamCloseQueuedOrInFlight(stream) === false);
-
-  const writer = stream._writer;
-  if (writer !== undefined && backpressure !== stream._backpressure) {
-    if (backpressure === true) {
-      defaultWriterReadyPromiseReset(writer);
-    } else {
-      assert(backpressure === false);
-
-      defaultWriterReadyPromiseResolve(writer);
-    }
-  }
-
-  stream._backpressure = backpressure;
-}
-
-class WritableStreamDefaultWriter {
-  constructor(stream) {
-    if (IsWritableStream(stream) === false) {
-      throw new TypeError('WritableStreamDefaultWriter can only be constructed with a WritableStream instance');
-    }
-    if (IsWritableStreamLocked(stream) === true) {
-      throw new TypeError('This stream has already been locked for exclusive writing by another writer');
-    }
-
-    this._ownerWritableStream = stream;
-    stream._writer = this;
-
-    const state = stream._state;
-
-    if (state === 'writable') {
-      if (WritableStreamCloseQueuedOrInFlight(stream) === false && stream._backpressure === true) {
-        defaultWriterReadyPromiseInitialize(this);
-      } else {
-        defaultWriterReadyPromiseInitializeAsResolved(this);
-      }
-
-      defaultWriterClosedPromiseInitialize(this);
-    } else if (state === 'erroring') {
-      defaultWriterReadyPromiseInitializeAsRejected(this, stream._storedError);
-      setPromiseIsHandledToTrue(this._readyPromise);
-      defaultWriterClosedPromiseInitialize(this);
-    } else if (state === 'closed') {
-      defaultWriterReadyPromiseInitializeAsResolved(this);
-      defaultWriterClosedPromiseInitializeAsResolved(this);
-    } else {
-      assert(state === 'errored');
-
-      const storedError = stream._storedError;
-      defaultWriterReadyPromiseInitializeAsRejected(this, storedError);
-      setPromiseIsHandledToTrue(this._readyPromise);
-      defaultWriterClosedPromiseInitializeAsRejected(this, storedError);
-      setPromiseIsHandledToTrue(this._closedPromise);
-    }
-  }
-
-  get closed() {
-    if (IsWritableStreamDefaultWriter(this) === false) {
-      return promiseRejectedWith(defaultWriterBrandCheckException('closed'));
-    }
-
-    return this._closedPromise;
-  }
-
-  get desiredSize() {
-    if (IsWritableStreamDefaultWriter(this) === false) {
-      throw defaultWriterBrandCheckException('desiredSize');
-    }
-
-    if (this._ownerWritableStream === undefined) {
-      throw defaultWriterLockException('desiredSize');
-    }
-
-    return WritableStreamDefaultWriterGetDesiredSize(this);
-  }
-
-  get ready() {
-    if (IsWritableStreamDefaultWriter(this) === false) {
-      return promiseRejectedWith(defaultWriterBrandCheckException('ready'));
-    }
-
-    return this._readyPromise;
-  }
-
-  abort(reason) {
-    if (IsWritableStreamDefaultWriter(this) === false) {
-      return promiseRejectedWith(defaultWriterBrandCheckException('abort'));
-    }
-
-    if (this._ownerWritableStream === undefined) {
-      return promiseRejectedWith(defaultWriterLockException('abort'));
-    }
-
-    return WritableStreamDefaultWriterAbort(this, reason);
-  }
-
-  close() {
-    if (IsWritableStreamDefaultWriter(this) === false) {
-      return promiseRejectedWith(defaultWriterBrandCheckException('close'));
-    }
-
-    const stream = this._ownerWritableStream;
-
-    if (stream === undefined) {
-      return promiseRejectedWith(defaultWriterLockException('close'));
-    }
-
-    if (WritableStreamCloseQueuedOrInFlight(stream) === true) {
-      return promiseRejectedWith(new TypeError('Cannot close an already-closing stream'));
-    }
-
-    return WritableStreamDefaultWriterClose(this);
-  }
-
-  releaseLock() {
-    if (IsWritableStreamDefaultWriter(this) === false) {
-      throw defaultWriterBrandCheckException('releaseLock');
-    }
-
-    const stream = this._ownerWritableStream;
-
-    if (stream === undefined) {
-      return;
-    }
-
-    assert(stream._writer !== undefined);
-
-    WritableStreamDefaultWriterRelease(this);
-  }
-
-  write(chunk) {
-    if (IsWritableStreamDefaultWriter(this) === false) {
-      return promiseRejectedWith(defaultWriterBrandCheckException('write'));
-    }
-
-    if (this._ownerWritableStream === undefined) {
-      return promiseRejectedWith(defaultWriterLockException('write to'));
-    }
-
-    return WritableStreamDefaultWriterWrite(this, chunk);
-  }
-}
-
-// Abstract operations for the WritableStreamDefaultWriter.
-
-function IsWritableStreamDefaultWriter(x) {
-  if (!typeIsObject(x)) {
-    return false;
-  }
-
-  if (!Object.prototype.hasOwnProperty.call(x, '_ownerWritableStream')) {
-    return false;
-  }
-
-  return true;
-}
-
-// A client of WritableStreamDefaultWriter may use these functions directly to bypass state check.
-
-function WritableStreamDefaultWriterAbort(writer, reason) {
-  const stream = writer._ownerWritableStream;
-
-  assert(stream !== undefined);
-
-  return WritableStreamAbort(stream, reason);
-}
-
-function WritableStreamDefaultWriterClose(writer) {
-  const stream = writer._ownerWritableStream;
-
-  assert(stream !== undefined);
-
-  return WritableStreamClose(stream);
-}
-
-function WritableStreamDefaultWriterCloseWithErrorPropagation(writer) {
-  const stream = writer._ownerWritableStream;
-
-  assert(stream !== undefined);
-
-  const state = stream._state;
-  if (WritableStreamCloseQueuedOrInFlight(stream) === true || state === 'closed') {
-    return promiseResolvedWith(undefined);
-  }
-
-  if (state === 'errored') {
-    return promiseRejectedWith(stream._storedError);
-  }
-
-  assert(state === 'writable' || state === 'erroring');
-
-  return WritableStreamDefaultWriterClose(writer);
-}
-
-function WritableStreamDefaultWriterEnsureClosedPromiseRejected(writer, error) {
-  if (writer._closedPromiseState === 'pending') {
-    defaultWriterClosedPromiseReject(writer, error);
-  } else {
-    defaultWriterClosedPromiseResetToRejected(writer, error);
-  }
-  setPromiseIsHandledToTrue(writer._closedPromise);
-}
-
-function WritableStreamDefaultWriterEnsureReadyPromiseRejected(writer, error) {
-  verbose('WritableStreamDefaultWriterEnsureReadyPromiseRejected(writer, %o)', error);
-  if (writer._readyPromiseState === 'pending') {
-    defaultWriterReadyPromiseReject(writer, error);
-  } else {
-    defaultWriterReadyPromiseResetToRejected(writer, error);
-  }
-  setPromiseIsHandledToTrue(writer._readyPromise);
-}
-
-function WritableStreamDefaultWriterGetDesiredSize(writer) {
-  const stream = writer._ownerWritableStream;
-  const state = stream._state;
-
-  if (state === 'errored' || state === 'erroring') {
-    return null;
-  }
-
-  if (state === 'closed') {
-    return 0;
-  }
-
-  return WritableStreamDefaultControllerGetDesiredSize(stream._writableStreamController);
-}
-
-function WritableStreamDefaultWriterRelease(writer) {
-  const stream = writer._ownerWritableStream;
-  assert(stream !== undefined);
-  assert(stream._writer === writer);
-
-  const releasedError = new TypeError(
-    'Writer was released and can no longer be used to monitor the stream\'s closedness');
-
-  WritableStreamDefaultWriterEnsureReadyPromiseRejected(writer, releasedError);
-
-  // The state transitions to "errored" before the sink abort() method runs, but the writer.closed promise is not
-  // rejected until afterwards. This means that simply testing state will not work.
-  WritableStreamDefaultWriterEnsureClosedPromiseRejected(writer, releasedError);
-
-  stream._writer = undefined;
-  writer._ownerWritableStream = undefined;
-}
-
-function WritableStreamDefaultWriterWrite(writer, chunk) {
-  const stream = writer._ownerWritableStream;
-
-  assert(stream !== undefined);
-
-  const controller = stream._writableStreamController;
-
-  const chunkSize = WritableStreamDefaultControllerGetChunkSize(controller, chunk);
-
-  if (stream !== writer._ownerWritableStream) {
-    return promiseRejectedWith(defaultWriterLockException('write to'));
-  }
-
-  const state = stream._state;
-  if (state === 'errored') {
-    return promiseRejectedWith(stream._storedError);
-  }
-  if (WritableStreamCloseQueuedOrInFlight(stream) === true || state === 'closed') {
-    return promiseRejectedWith(new TypeError('The stream is closing or closed and cannot be written to'));
-  }
-  if (state === 'erroring') {
-    return promiseRejectedWith(stream._storedError);
-  }
-
-  assert(state === 'writable');
-
-  const promise = WritableStreamAddWriteRequest(stream);
-
-  WritableStreamDefaultControllerWrite(controller, chunk, chunkSize);
-
-  return promise;
-}
-
-class WritableStreamDefaultController {
-  constructor() {
-    throw new TypeError('WritableStreamDefaultController cannot be constructed explicitly');
-  }
-
-  error(e) {
-    if (IsWritableStreamDefaultController(this) === false) {
-      throw new TypeError(
-        'WritableStreamDefaultController.prototype.error can only be used on a WritableStreamDefaultController');
-    }
-    const state = this._controlledWritableStream._state;
-    if (state !== 'writable') {
-      // The stream is closed, errored or will be soon. The sink can't do anything useful if it gets an error here, so
-      // just treat it as a no-op.
-      return;
-    }
-
-    WritableStreamDefaultControllerError(this, e);
-  }
-
-  [AbortSteps](reason) {
-    const result = this._abortAlgorithm(reason);
-    WritableStreamDefaultControllerClearAlgorithms(this);
-    return result;
-  }
-
-  [ErrorSteps]() {
-    ResetQueue(this);
-  }
-}
-
-// Abstract operations implementing interface required by the WritableStream.
-
-function IsWritableStreamDefaultController(x) {
-  if (!typeIsObject(x)) {
-    return false;
-  }
-
-  if (!Object.prototype.hasOwnProperty.call(x, '_controlledWritableStream')) {
-    return false;
-  }
-
-  return true;
-}
-
-function SetUpWritableStreamDefaultController(stream, controller, startAlgorithm, writeAlgorithm, closeAlgorithm,
-                                              abortAlgorithm, highWaterMark, sizeAlgorithm) {
-  assert(IsWritableStream(stream) === true);
-  assert(stream._writableStreamController === undefined);
-
-  controller._controlledWritableStream = stream;
-  stream._writableStreamController = controller;
-
-  // Need to set the slots so that the assert doesn't fire. In the spec the slots already exist implicitly.
-  controller._queue = undefined;
-  controller._queueTotalSize = undefined;
-  ResetQueue(controller);
-
-  controller._started = false;
-
-  controller._strategySizeAlgorithm = sizeAlgorithm;
-  controller._strategyHWM = highWaterMark;
-
-  controller._writeAlgorithm = writeAlgorithm;
-  controller._closeAlgorithm = closeAlgorithm;
-  controller._abortAlgorithm = abortAlgorithm;
-
-  const backpressure = WritableStreamDefaultControllerGetBackpressure(controller);
-  WritableStreamUpdateBackpressure(stream, backpressure);
-
-  const startResult = startAlgorithm();
-  const startPromise = promiseResolvedWith(startResult);
-  uponPromise(
-    startPromise,
-    () => {
-      assert(stream._state === 'writable' || stream._state === 'erroring');
-      controller._started = true;
-      WritableStreamDefaultControllerAdvanceQueueIfNeeded(controller);
-    },
-    r => {
-      assert(stream._state === 'writable' || stream._state === 'erroring');
-      controller._started = true;
-      WritableStreamDealWithRejection(stream, r);
-    }
-  );
-}
-
-function SetUpWritableStreamDefaultControllerFromUnderlyingSink(stream, underlyingSink, highWaterMark, sizeAlgorithm) {
-  assert(underlyingSink !== undefined);
-
-  const controller = Object.create(WritableStreamDefaultController.prototype);
-
-  function startAlgorithm() {
-    return InvokeOrNoop(underlyingSink, 'start', [controller]);
-  }
-
-  const writeAlgorithm = CreateAlgorithmFromUnderlyingMethod(underlyingSink, 'write', 1, [controller]);
-  const closeAlgorithm = CreateAlgorithmFromUnderlyingMethod(underlyingSink, 'close', 0, []);
-  const abortAlgorithm = CreateAlgorithmFromUnderlyingMethod(underlyingSink, 'abort', 1, []);
-
-  SetUpWritableStreamDefaultController(stream, controller, startAlgorithm, writeAlgorithm, closeAlgorithm,
-                                       abortAlgorithm, highWaterMark, sizeAlgorithm);
-}
-
-// ClearAlgorithms may be called twice. Erroring the same stream in multiple ways will often result in redundant calls.
-function WritableStreamDefaultControllerClearAlgorithms(controller) {
-  controller._writeAlgorithm = undefined;
-  controller._closeAlgorithm = undefined;
-  controller._abortAlgorithm = undefined;
-  controller._strategySizeAlgorithm = undefined;
-}
-
-function WritableStreamDefaultControllerClose(controller) {
-  EnqueueValueWithSize(controller, 'close', 0);
-  WritableStreamDefaultControllerAdvanceQueueIfNeeded(controller);
-}
-
-function WritableStreamDefaultControllerGetChunkSize(controller, chunk) {
-  try {
-    return controller._strategySizeAlgorithm(chunk);
-  } catch (chunkSizeE) {
-    WritableStreamDefaultControllerErrorIfNeeded(controller, chunkSizeE);
-    return 1;
-  }
-}
-
-function WritableStreamDefaultControllerGetDesiredSize(controller) {
-  return controller._strategyHWM - controller._queueTotalSize;
-}
-
-function WritableStreamDefaultControllerWrite(controller, chunk, chunkSize) {
-  const writeRecord = { chunk };
-
-  try {
-    EnqueueValueWithSize(controller, writeRecord, chunkSize);
-  } catch (enqueueE) {
-    WritableStreamDefaultControllerErrorIfNeeded(controller, enqueueE);
-    return;
-  }
-
-  const stream = controller._controlledWritableStream;
-  if (WritableStreamCloseQueuedOrInFlight(stream) === false && stream._state === 'writable') {
-    const backpressure = WritableStreamDefaultControllerGetBackpressure(controller);
-    WritableStreamUpdateBackpressure(stream, backpressure);
-  }
-
-  WritableStreamDefaultControllerAdvanceQueueIfNeeded(controller);
-}
-
-// Abstract operations for the WritableStreamDefaultController.
-
-function WritableStreamDefaultControllerAdvanceQueueIfNeeded(controller) {
-  verbose('WritableStreamDefaultControllerAdvanceQueueIfNeeded()');
-  const stream = controller._controlledWritableStream;
-
-  if (controller._started === false) {
-    return;
-  }
-
-  if (stream._inFlightWriteRequest !== undefined) {
-    return;
-  }
-
-  const state = stream._state;
-  assert(state !== 'closed' && state !== 'errored');
-  if (state === 'erroring') {
-    WritableStreamFinishErroring(stream);
-    return;
-  }
-
-  if (controller._queue.length === 0) {
-    return;
-  }
-
-  const writeRecord = PeekQueueValue(controller);
-  if (writeRecord === 'close') {
-    WritableStreamDefaultControllerProcessClose(controller);
-  } else {
-    WritableStreamDefaultControllerProcessWrite(controller, writeRecord.chunk);
-  }
-}
-
-function WritableStreamDefaultControllerErrorIfNeeded(controller, error) {
-  if (controller._controlledWritableStream._state === 'writable') {
-    WritableStreamDefaultControllerError(controller, error);
-  }
-}
-
-function WritableStreamDefaultControllerProcessClose(controller) {
-  const stream = controller._controlledWritableStream;
-
-  WritableStreamMarkCloseRequestInFlight(stream);
-
-  DequeueValue(controller);
-  assert(controller._queue.length === 0);
-
-  const sinkClosePromise = controller._closeAlgorithm();
-  WritableStreamDefaultControllerClearAlgorithms(controller);
-  uponPromise(
-    sinkClosePromise,
-    () => {
-      WritableStreamFinishInFlightClose(stream);
-    },
-    reason => {
-      WritableStreamFinishInFlightCloseWithError(stream, reason);
-    }
-  );
-}
-
-function WritableStreamDefaultControllerProcessWrite(controller, chunk) {
-  const stream = controller._controlledWritableStream;
-
-  WritableStreamMarkFirstWriteRequestInFlight(stream);
-
-  const sinkWritePromise = controller._writeAlgorithm(chunk);
-  uponPromise(
-    sinkWritePromise,
-    () => {
-      WritableStreamFinishInFlightWrite(stream);
-
-      const state = stream._state;
-      assert(state === 'writable' || state === 'erroring');
-
-      DequeueValue(controller);
-
-      if (WritableStreamCloseQueuedOrInFlight(stream) === false && state === 'writable') {
-        const backpressure = WritableStreamDefaultControllerGetBackpressure(controller);
-        WritableStreamUpdateBackpressure(stream, backpressure);
-      }
-
-      WritableStreamDefaultControllerAdvanceQueueIfNeeded(controller);
-    },
-    reason => {
-      if (stream._state === 'writable') {
-        WritableStreamDefaultControllerClearAlgorithms(controller);
-      }
-      WritableStreamFinishInFlightWriteWithError(stream, reason);
-    }
-  );
-}
-
-function WritableStreamDefaultControllerGetBackpressure(controller) {
-  const desiredSize = WritableStreamDefaultControllerGetDesiredSize(controller);
-  return desiredSize <= 0;
-}
-
-// A client of WritableStreamDefaultController may use these functions directly to bypass state check.
-
-function WritableStreamDefaultControllerError(controller, error) {
-  const stream = controller._controlledWritableStream;
-
-  assert(stream._state === 'writable');
-
-  WritableStreamDefaultControllerClearAlgorithms(controller);
-  WritableStreamStartErroring(stream, error);
-}
-
-// Helper functions for the WritableStream.
-
-function streamBrandCheckException(name) {
-  return new TypeError(`WritableStream.prototype.${name} can only be used on a WritableStream`);
-}
-
-// Helper functions for the WritableStreamDefaultWriter.
-
-function defaultWriterBrandCheckException(name) {
-  return new TypeError(
-    `WritableStreamDefaultWriter.prototype.${name} can only be used on a WritableStreamDefaultWriter`);
-}
-
-function defaultWriterLockException(name) {
-  return new TypeError('Cannot ' + name + ' a stream using a released writer');
-}
-
-function defaultWriterClosedPromiseInitialize(writer) {
-  writer._closedPromise = newPromise((resolve, reject) => {
-    writer._closedPromise_resolve = resolve;
-    writer._closedPromise_reject = reject;
-    writer._closedPromiseState = 'pending';
-  });
-}
-
-function defaultWriterClosedPromiseInitializeAsRejected(writer, reason) {
-  writer._closedPromise = promiseRejectedWith(reason);
-  writer._closedPromise_resolve = undefined;
-  writer._closedPromise_reject = undefined;
-  writer._closedPromiseState = 'rejected';
-}
-
-function defaultWriterClosedPromiseInitializeAsResolved(writer) {
-  writer._closedPromise = promiseResolvedWith(undefined);
-  writer._closedPromise_resolve = undefined;
-  writer._closedPromise_reject = undefined;
-  writer._closedPromiseState = 'resolved';
-}
-
-function defaultWriterClosedPromiseReject(writer, reason) {
-  assert(writer._closedPromise_resolve !== undefined);
-  assert(writer._closedPromise_reject !== undefined);
-  assert(writer._closedPromiseState === 'pending');
-
-  writer._closedPromise_reject(reason);
-  writer._closedPromise_resolve = undefined;
-  writer._closedPromise_reject = undefined;
-  writer._closedPromiseState = 'rejected';
-}
-
-function defaultWriterClosedPromiseResetToRejected(writer, reason) {
-  assert(writer._closedPromise_resolve === undefined);
-  assert(writer._closedPromise_reject === undefined);
-  assert(writer._closedPromiseState !== 'pending');
-
-  writer._closedPromise = promiseRejectedWith(reason);
-  writer._closedPromiseState = 'rejected';
-}
-
-function defaultWriterClosedPromiseResolve(writer) {
-  assert(writer._closedPromise_resolve !== undefined);
-  assert(writer._closedPromise_reject !== undefined);
-  assert(writer._closedPromiseState === 'pending');
-
-  writer._closedPromise_resolve(undefined);
-  writer._closedPromise_resolve = undefined;
-  writer._closedPromise_reject = undefined;
-  writer._closedPromiseState = 'resolved';
-}
-
-function defaultWriterReadyPromiseInitialize(writer) {
-  verbose('defaultWriterReadyPromiseInitialize()');
-  writer._readyPromise = newPromise((resolve, reject) => {
-    writer._readyPromise_resolve = resolve;
-    writer._readyPromise_reject = reject;
-  });
-  writer._readyPromiseState = 'pending';
-}
-
-function defaultWriterReadyPromiseInitializeAsRejected(writer, reason) {
-  verbose('defaultWriterReadyPromiseInitializeAsRejected(writer, %o)', reason);
-  writer._readyPromise = promiseRejectedWith(reason);
-  writer._readyPromise_resolve = undefined;
-  writer._readyPromise_reject = undefined;
-  writer._readyPromiseState = 'rejected';
-}
-
-function defaultWriterReadyPromiseInitializeAsResolved(writer) {
-  verbose('defaultWriterReadyPromiseInitializeAsResolved()');
-  writer._readyPromise = promiseResolvedWith(undefined);
-  writer._readyPromise_resolve = undefined;
-  writer._readyPromise_reject = undefined;
-  writer._readyPromiseState = 'fulfilled';
-}
-
-function defaultWriterReadyPromiseReject(writer, reason) {
-  verbose('defaultWriterReadyPromiseReject(writer, %o)', reason);
-  assert(writer._readyPromise_resolve !== undefined);
-  assert(writer._readyPromise_reject !== undefined);
-
-  writer._readyPromise_reject(reason);
-  writer._readyPromise_resolve = undefined;
-  writer._readyPromise_reject = undefined;
-  writer._readyPromiseState = 'rejected';
-}
-
-function defaultWriterReadyPromiseReset(writer) {
-  verbose('defaultWriterReadyPromiseReset()');
-  assert(writer._readyPromise_resolve === undefined);
-  assert(writer._readyPromise_reject === undefined);
-
-  writer._readyPromise = newPromise((resolve, reject) => {
-    writer._readyPromise_resolve = resolve;
-    writer._readyPromise_reject = reject;
-  });
-  writer._readyPromiseState = 'pending';
-}
-
-function defaultWriterReadyPromiseResetToRejected(writer, reason) {
-  verbose('defaultWriterReadyPromiseResetToRejected(writer, %o)', reason);
-  assert(writer._readyPromise_resolve === undefined);
-  assert(writer._readyPromise_reject === undefined);
-
-  writer._readyPromise = promiseRejectedWith(reason);
-  writer._readyPromiseState = 'rejected';
-}
-
-function defaultWriterReadyPromiseResolve(writer) {
-  verbose('defaultWriterReadyPromiseResolve()');
-  assert(writer._readyPromise_resolve !== undefined);
-  assert(writer._readyPromise_reject !== undefined);
-
-  writer._readyPromise_resolve(undefined);
-  writer._readyPromise_resolve = undefined;
-  writer._readyPromise_reject = undefined;
-  writer._readyPromiseState = 'fulfilled';
-}
